requirements.txt
Flask
psycopg
openai
python-dotenv
Flask-Cors
pandas